Add Levels singleton to access and save levels.
Levels singleton is added. It allows iteration through an array of
levels. Classes can load levels from the paths in the array, and load
save information, as well as receive level information from config
files. This class was created as a stepping stone for creating the
level select and start game actions in the user interface.