Fix text align change after disabling richtext.
When rich text was disabled, all the formatting was being cleared from
it, including the centering. This was because calls to the set_text()
function had 5 arguments instead of 6. The animate argument was being
omitted, so the value of "use rich text" was being applied to whether or
not to center the text.