Fix to broken depth system.
* Also includes fixes for runtime crashes.

The depth controllers set_collision_layer() and set_collision_mask()
functions were not working as intended. The idea was that up to
COLLISION_OFFSET layers could be in a depth layer, and switching layers
would appropriately shift a Node's mask and layer. Instead, it was
taking the power of the layer, which was destructive and caused layers
to be incorrect. Now, the layer and mask are calculated by shifting the
starting layer/mask (layer and mask at depth 0) over by depth *
COLLISION_OFFSET, which seems to work.