Fix AbilitiesDisplay showing multiples copies.
Each time _ready() was called in the level, the abilities display was
duplicating itself. This was caused by the abilities ready signal being
emitted multiple times in _ready(). For now, I've changed it in
AbilitiesDisplay to only accept the signal once. If it's ever planned
that abilities can be gained during a level, then it might be better to
add a reset function instead.