Fix incorrect size change methods for health bar.
The shink and grow methods for the health bar were modifying the
underlying list in the wrong way. The tail is actually index 0 in array
storage, and the head is n - 1.

To remove elements, elements from the
tail and up should get removed, and the cutoff point needs to also be
removed for a tail.

To add elements, the previous tail must be replaced, and the new
elements should be inserted at the front of the list.

Also changed were the _test_shrink() and _test_grow() methods, which now
have damaged segments in certain areas to make it more obvious how the
list is changed when the health bar grows or shrinks.