Update obstacle.gd to remove useless Node2D.
Node2D didn't add anything that couldn't be added by having StaticBody2D
as the parent, so it was removed. This works out better because
SmartShape2D uses StaticBody2D as a parent as well. Once the drawing
code is cleaned out, obstacle.gd should be ready to be added to
SmartShape2D nodes.