Add enemy sprites and fix some AI bugs.
Enemy is now fully animated.

Fixed a bug where the end_<blank>_attack() functions in basic_enemy.gd
were calling fsm.pop(). This was causing the death state to be popped
sometimes, which would resurrect the enemy. This resurrection would make
the enemy unkillable, since its health was below 0, and it would be
entirely unable to take any form of damage.