Add gore layer to worm head node.
Added a layer that can be overlaid on the worm to show blood effects.
Currently, it is not linked to any actions. The modulate is set to (1,
1, 1, 0), so it's fully transparent.

To make it work in game, I think making the blood effects come in
progressively by tweening the modulate update by 0.25 of the max on each
enemy eaten would look good.