Fix problem with enemy dying during attacks.
The "wakeup" problem of an enemy dying while attacking, then waking up
on animation end should be corrected. When the enemy leaves the attack
state, the signal handler is disconnected.

For the melee attack, the call that disables the melee attack collision
should also be called when an enemy dies now.