Change bite from one action to two.
Previously, the bite ability was a very long animation that was hard to
time since the last part of it was so sudden. This made it hard to aim.
Considering its use as the primary attack in the game, I decided to
split it up into two actions to make it easier to aim.

In doing so, I removed a lot of the animation callbacks from
head_kinematic, and put them directly in bite. Not every ability has an
animation, so this behaviour should be on a per ability basis.