Fix camera and lag spike on bite enemy.
When biting an enemy, the camera would just far away to the origin for a
frame before the new tail segment was moved to the correct position.
This no longer happens, because the new tail segment is moved to the
correct position immediately when it spawns.

A lag spike was occurring on bites because a lot of signals were getting
sent out all at once when healing the player. A small delay of 100ms has
been introduced between the heals, which eases the load on the signal
handler system. A better solution should be introduced eventually, where
a damage on many segments at the same time can be processed with a
single signal. This would entail adding a batch update to HealthBar,
SpawnKinematic, and anything else that responds when the player takes
damage.