Fix slider display. Work on change keybinds.
VSlider and HSlider were using the same borders, which was causing a
display issue for the hsliders. Each type of slider has its own theme
now.

I started making the keybinds menu for the settings screen. I go as
far as connecting all the pressed and mouse_entered signals of the
buttons. The next step would be to create a popup menu that prompts the
user to press any key. Ideally, the menu should provide the option to
change its title and message, so we can just use a single instance for
each control.