Add change key bind menu.
Created a popup menu for changing key bindings. It is styled in the same
way as the settings menu is, and is meant for changing the bindings
associated with actions.

A single instance of a menu can be resused, by calling show_for with the
action needed. The user can cancel the menu by clicking off it, or
pressing escape. OK must be pressed for the new binding to take.