Add singletons to apply Configuration changes.
Started to add a new system for applying saved Configuration values to
the game at boot. A singleton is created for each section of
Configuration. The singleton contains methods for each key in
Configuration. The methods apply a single Configuration value to the
game, reading it directly from Configuration. This system should keep
the responsibility of applying configuration in one place. Other nodes
should only ever have to write to Configuration or call
<Section>ConfigLoader.apply_<setting>() to make the new value
immediately take effect.