Add default resolution when out of level.
The user's resolution setting is now only used when the user is out the
menus. The reason for this is so that the user can close the application
and relaunch it if they pick a weird resolution. Since the UI is also
affected by the resolution setting, this is necessary because the text
resolution changes too, making it really hard to read.