Add viewport test for better reslxn options.
Viewport test proof of concept was added that shows a better way to do
downscaling the resolution. Game elements should be placed in a
viewport, and UI should rest on top of the viewport. Resolution can be
lowered by increasing stretch scale, and zooming out the camera by the
same factor.