Optimize enemy player detection code.
Each frame, enemies would look for the player and update their movement.
This was having a noticeable impact on performance. Counters have been
added to limit the amount of times an enemy will update its tracking and
movement to improve performance. The distance function has been replaced
in most enemy states to use Manhattan distance instead of the sqrt
formula.