Add interface for saving player data.
I wrote a singleton that can be used to read and write from the player's
save. Most methods return either OK, or an error constant. It is not yet
set up, but this should probably be a singleton or that loads once at
the start of the game, and writes once at the end of the game.