Prevent locked skins being saved to current skin.
The player was able to save a locked skin as their current skin in the
skin select by selecting a "???" skin and exiting from skin select. The
game was not checking if the skin was unlocked before saving the current
skin, so the player could equip locked skins through this approach.