Read from PlayerSave to modify UI, load skins.
- Level select uses the save file to load information about levels
including the current progress of the level. At this time, through level
select levels can be played out of order, but this may be changed in the
future.
- Title screen uses the level progress to determine what level to launch
when the player clicks the start game button, and will show "Start Game"
instead of "Continue" if there is no level progress.
- Skin select will "black" out skins that have not yet been unlocked.