Disable some graphical effects wrt configuration.
Player and enemy blood appearing on the player model can be disabled
altogether in order to some some processing time when the player is
damaged.

Support for reducing the number of particles from the particle emitters
has been partially added. An extension for CPUParticles2D that throttles
that amount of particles was created, and should be distributed around
to each CPUParticles instance.